Create targetjournals.md
for discussion about which journal the final article should be submitted to.

dwald's original question in README.md: *Do we have a target journal in mind? Can we add a note about this in the article document or maybe also ask that collaboratively through GitHub or IdeaScale?*

JustinLongo's response: Good question. I'm going to move this to an Issue. Once we have a number of candidate journals, we can think about moving it to IdeaScale.